+++ Source/Examples/Xamarin.Forms/IssueDemos/IssueDemos.iOS/AppDelegate.cs
@@ -23,7 +23,7 @@ namespace IssueDemos.iOS
         public override bool FinishedLaunching(UIApplication app, NSDictionary options)
         {
             global::Xamarin.Forms.Forms.Init();
-            OxyPlot.Xamarin.Forms.Platform.iOS.Forms.Init();
+            OxyPlot.Xamarin.Forms.Platform.iOS.PlotViewRenderer.Init();
             LoadApplication(new App());
 
             return base.FinishedLaunching(app, options);