+++ Source/Examples/Xamarin.Forms/SimpleDemo2/SimpleDemo2.Droid/MainActivity.cs
@@ -17,7 +17,7 @@ namespace SimpleDemo2.Droid
 			base.OnCreate (bundle);
 
 			global::Xamarin.Forms.Forms.Init (this, bundle);
-		    OxyPlot.Xamarin.Forms.Platform.Android.Forms.Init();
+		    OxyPlot.Xamarin.Forms.Platform.Android.PlotViewRenderer.Init();
 			LoadApplication (new SimpleDemo2.App ());
 		}
 	}