+++ Source/OxyPlot.Xamarin.Forms.Platform.iOS.Classic/Forms.cs
@@ -9,6 +9,8 @@
 
 namespace OxyPlot.Xamarin.Forms.Platform.iOS.Classic
 {
+    using System;
+
     /// <summary>
     /// Initializes OxyPlot renderers for use with Xamarin.Forms.
     /// </summary>
@@ -17,12 +19,11 @@ namespace OxyPlot.Xamarin.Forms.Platform.iOS.Classic
         /// <summary>
         /// Initializes OxyPlot for Xamarin.Forms.
         /// </summary>
-        /// <remarks>This method must be called before Forms.Init().</remarks>
+        /// <remarks>This method must be called before a <see cref="T:PlotView" /> is used.</remarks>
+        [Obsolete("Use PlotViewRenderer.Init() instead.")]
         public static void Init()
         {
-            // Just bring this assembly into the current appdomain.
-            // Forms.Init() should now find it!
-            PlotView.IsRendererInitialized = true;
+            PlotViewRenderer.Init();
         }
     }
 }